made lat and lon doubles

+++ src/main/java/jais/messages/BaseStationReport.java
@@ -39,8 +39,8 @@ public class BaseStationReport extends AISMessageBase {
     private int minute;
     private int second;
     private boolean accurate;
-    private float lon;
-    private float lat;
+    private double lon;
+    private double lat;
     private EPFDFixType epfd;
     private boolean raim;
     private int radio;
